// watch-load/generated/prisma/models/HeartMeasurement.ts
export type HeartMeasurementCreateManyInput = {
  id?: string
  signalId: string
  deviceId: string
  heartRate: number
  afib: $Enums.Afib
  signal?: Prisma.HeartMeasurementCreatesignalInput | number[]
  samplingFrequency: number
  timestamp: Date | string
  modified: Date | string
  trialsId?: string | null
  locationId?: string | null
  createdAt?: Date | string
  updatedAt?: Date | string
  workspaceId: string
}
// ...
export type HeartMeasurementCreatesignalInput = {
  set: number[]
}